requirements.txt
numpy==1.24.3
opencv_python==4.8.0.74
pandas==2.0.3
Pillow==10.0.0
Pillow==10.3.0
scikit_learn==1.3.0
tensorflow==2.13.0
tqdm==4.66.0
ultralytics==8.1.24
